requirements.txt
amqp==2.5.2
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.0
Django==2.2.12
django-celery-beat==2.0.0
django-filter==2.2.0
django-timezone-field==4.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
kombu==4.6.8
more-itertools==7.2.0
mysql==0.0.2
mysqlclient==1.4.6
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycurl==7.43.0.3
Pygments==2.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.2
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.2
urllib3==1.25.9
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.7
wget==3.2
zipp==1.2.0
